allow to override the URL provided in the WSDL document
The WSDL document provided by the TICKeos API contains the wrong
hostnames if it is downloaded from the development system and requests
should go to the development system. For example:

The service section is:
<wsdl:service name="kombiticket">
  <wsdl:port binding="tns:kombiticketSOAP" name="kombiticketSOAP">
    <soap:address location="https://tickeos.de/service.php/kombiticket/1_5"/>
  </wsdl:port>
</wsdl:service>

However, the correct URL would have a differnt hostname. The option
alternative_address allows to override this.

The workaround might fail with later versions of Zeep because it relies
on the non-public API.